Import gtk2 style from qtbase
We are going to drop gtk2 support from qtbase, so move that style plugin
from qtbase to here.

As a bonus, we can now link directly to the libraries we use, instead
of loading them dynamically (in qtbase this style was part of QtWidgets
library, and we didn't want it to link against gtk2).

Change-Id: I2fc3325fbc0a9f2588b99e94aa906eebbe76d45c
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>