Ensure the right color is used for drawing the label text in a combobox
After QWindowsStyle was changed to use a different color from the palette
for the CE_ComboBoxLabel case it caused it to override the color used by
the plastique style. By going right to QCommonStyle we bypass the changes
done by the Windows style and thus ensure it is drawn correctly.

Change-Id: Ic1398afc16e2498f25287f55526301fb7f17731c
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>