Add missing PLUGIN_CLASS_NAMEs
It will create a separate imported target for each plugin
instead of a single Qt5Widgets_.cmake for an arbirary one.
An imported target can be used e.g. for reading the
location property of a plugin.

Also change PLUGIN_CLASS_NAME for the bb10style plugin to
match the class name.

Change-Id: I0bdf98c4d47ee9e33611cb35a2f05f41e1948c7f
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>